requirements.txt
PyYAML
stix2
json
os
glob
urllib